Initial working block
Get this working by cloning obs-studio and then cloning this within
\plugins\

+++ src/audio-capture-helper.cpp
@@ -130,8 +130,8 @@ void AudioCaptureHelper::Capture()
 
 	bool shutdown = false;
 	while (!shutdown) {
-		auto event_id = WaitForMultipleObjects(events.size(), events[0].addressof(), FALSE,
-						       INFINITE);
+		auto event_id = WaitForMultipleObjects((DWORD)events.size(), events[0].addressof(),
+						       FALSE, INFINITE);
 
 		switch (event_id) {
 		case HelperEvents::PacketReady:
